airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
pyarrow
psycopg2-binary
elasticsearch
apache-airflow-providers-google